Created ability to have battles with multiple pokemon
After one of your pokemon faints, it is removed from the list and you are asked to enter the name of you next pokemon.  This goes until the list is empty.  With the opponent, after their pokemon faints they send out another random pokemon from the list.  This works but the turn doesn't stop if a pokemon faints.  The incoming pokemon then uses the attack that is then performed by that pokemon.